requirements.txt
aiogram==3.7.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.0
redis==5.0.1
alembic==1.13.1